# docs/swagger.yaml
basePath: /api
host: localhost:8080
info:
  contact: {}
  description: API for accessing course information
  title: Course API
  version: "1.0"
paths:
  /courses:
    get:
      consumes:
      - application/json
      description: Get a list of all courses, with optional search and pagination
      parameters:
      - description: Search courses by department code prefix (e.g., 'CSC' returns
          CSC110, CSC111, CSC320, etc.)
        in: query
        name: search
        type: string
      - default: 1
        description: 'Page number for pagination (default: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: 'Number of results per page (default: 20)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Course not found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Search the catalog of UVic courses
      tags:
      - courses
  /courses/info:
    get:
      consumes:
      - application/json
      description: Get general information about available courses
      parameters:
      - description: Program ID (in uppercase)
        in: query
        name: pid
        type: string
      - description: Course code (in uppercase)
        in: query
        name: course
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with course information
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Course not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Error when neither pid nor course is provided
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get course information
      tags:
      - courses
  /courses/sections/{term}/{course}:
    get:
      consumes:
      - application/json
      description: Get sections for a specific course in a specific term
      parameters:
      - description: Term ID (e.g., 202101)
        in: path
        name: term
        required: true
        type: string
      - description: Course ID (e.g., CSC110)
        in: path
        name: course
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with sections data or empty array if no
            sections found
          schema:
            type: object
        "500":
          description: Error when sections count is invalid
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get course sections
      tags:
      - courses
  /outline/{term}/{course}:
    get:
      consumes:
      - application/json
      description: Get the outline for a specific course in a specific term
      parameters:
      - description: Term ID
        in: path
        name: term
        required: true
        type: string
      - description: Course ID
        in: path
        name: course
        required: true
        type: string
      - default: false
        description: Include unpublished outlines if true
        in: query
        name: unpublished
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with outline link and validity
          schema:
            properties:
              isValid:
                type: boolean
              link:
                type: string
            type: object
        "500":
          description: Server error when fetching outline
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get course outline
      tags:
      - courses
swagger: "2.0"